refactor app startup code
